% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitpopsurv-package.R
\docType{package}
\name{splitpopsurv-package}
\alias{splitpopsurv}
\alias{splitpopsurv-package}
\title{splitpopsurv: Split-Population (Cure / Mover-Stayer) Survival Models}
\description{
Maximum-likelihood estimation of split-population survival models, which
combine an accelerated failure-time regression for event timing among
"movers" with a logistic regression on the probability of belonging to
the immune "stayer" population. See Schmidt & Witte (1989) and Yamaguchi
(1992, 1998) for the underlying theory, and the package README for a full
manual including a likelihood correction relative to the Stata `ml`
programs this package translates.
}
\section{Main functions}{

[fit_splitpop_loglogistic()], [fit_splitpop_weibull()],
[fit_splitpop_lognormal()], [fit_splitpop_gamma()],
[fit_splitpop_ggamma()].
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nobifukuda/splitpopsurv}
  \item \url{https://nobifukuda.github.io/splitpopsurv/}
  \item Report bugs at \url{https://github.com/nobifukuda/splitpopsurv/issues}
}

}
\author{
\strong{Maintainer}: Nobutaka Fukuda \email{nobutaka.fukuda@tohoku.ac.jp}

Authors:
\itemize{
  \item Nobutaka Fukuda \email{nobutaka.fukuda@tohoku.ac.jp}
}

}
\keyword{internal}
